+++ src/torchpme/calculators/calculator.py
@@ -25,15 +25,12 @@ class Calculator(torch.nn.Module):
         well as the parameters that determine the behavior of the potential itself.
     :param full_neighbor_list: parameter indicating whether the neighbor information
         will come from a full (True) or half (False, default) neighbor list.
-    :param prefactor: electrostatics prefactor; see :ref:`prefactors` for details and
-        common values.
     """
 
     def __init__(
         self,
         potential: Potential,
         full_neighbor_list: bool = False,
-        prefactor: float = 1.0,
     ):
         super().__init__()
 
@@ -44,7 +41,6 @@ class Calculator(torch.nn.Module):
 
         self.potential = potential
         self.full_neighbor_list = full_neighbor_list
-        self.prefactor = prefactor
 
     def _compute_rspace(
         self,
@@ -181,7 +177,7 @@ class Calculator(torch.nn.Module):
         )
 
         if self.potential.smearing is None:
-            return self.prefactor * potential_sr
+            return potential_sr
         # Compute long-range (LR) part using a Fourier / reciprocal space sum
         potential_lr = self._compute_kspace(
             charges=charges,
@@ -192,4 +188,4 @@ class Calculator(torch.nn.Module):
             node_mask=node_mask,
         )
 
-        return self.prefactor * (potential_sr + potential_lr)
+        return potential_sr + potential_lr