+++ docs/src/references/changelog.rst
@@ -23,10 +23,17 @@ changelog <https://keepachangelog.com/en/1.1.0/>`_ format. This project follows
 
 `Unreleased <https://github.com/lab-cosmo/torch-pme/>`_
 -------------------------------------------------------
+
 Added
 #####
 
 * Add support for batched calculations
+
+Changed
+#######
+
+* ``prefactor`` is now a parameter of the ``Potential`` classes instead of the
+  ``Calculator`` classes.
 
 
 `Version 0.3.2 <https://github.com/lab-cosmo/torch-pme/releases/tag/v0.3.2>`_ - 2025-10-07